IT and unit tests sepration. Minor clean code
