+++ src/test/java/com/project/manager/demo/DemoApplicationTests.java
@@ -1,8 +1,6 @@
 package com.project.manager.demo;
 
-import com.project.manager.demo.repository.UserRepository;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
